requirements.txt
django==4.2.2
pytest-django==4.5.1
factory-boy==3.2.1
djangorestframework==3.14.0
django-cors-headers==4.2.0